+++ udata/__init__.py
@@ -4,5 +4,5 @@
 udata
 '''
 
-__version__ = '6.1.7'
+__version__ = '6.1.8.dev'
 __description__ = 'Open data portal'